+++ cishouseholds/pipeline/job_transformations.py
@@ -5,60 +5,30 @@ from pyspark.sql import Window
 from cishouseholds.derive import assign_work_patient_facing_now
 from cishouseholds.derive import assign_work_person_facing_now
 from cishouseholds.derive import assign_work_status_group
-from cishouseholds.derive import flag_records_for_childcare_v0_rules
-from cishouseholds.derive import flag_records_for_childcare_v1_rules
-from cishouseholds.derive import flag_records_for_childcare_v2_b_rules
-from cishouseholds.derive import flag_records_for_college_v0_rules
-from cishouseholds.derive import flag_records_for_college_v1_rules
-from cishouseholds.derive import flag_records_for_college_v2_rules
-from cishouseholds.derive import flag_records_for_furlough_rules_v0
-from cishouseholds.derive import flag_records_for_furlough_rules_v1_a
-from cishouseholds.derive import flag_records_for_furlough_rules_v1_b
-from cishouseholds.derive import flag_records_for_furlough_rules_v2_a
-from cishouseholds.derive import flag_records_for_furlough_rules_v2_b
-from cishouseholds.derive import flag_records_for_not_working_rules_v0
-from cishouseholds.derive import flag_records_for_not_working_rules_v1_a
-from cishouseholds.derive import flag_records_for_not_working_rules_v1_b
-from cishouseholds.derive import flag_records_for_not_working_rules_v2_a
-from cishouseholds.derive import flag_records_for_not_working_rules_v2_b
-from cishouseholds.derive import flag_records_for_retired_rules
-from cishouseholds.derive import flag_records_for_school_v2_rules
-from cishouseholds.derive import flag_records_for_self_employed_rules_v0
-from cishouseholds.derive import flag_records_for_self_employed_rules_v1_a
-from cishouseholds.derive import flag_records_for_self_employed_rules_v1_b
-from cishouseholds.derive import flag_records_for_self_employed_rules_v2_a
-from cishouseholds.derive import flag_records_for_self_employed_rules_v2_b
-from cishouseholds.derive import flag_records_for_uni_v0_rules
-from cishouseholds.derive import flag_records_for_uni_v1_rules
-from cishouseholds.derive import flag_records_for_uni_v2_rules
-from cishouseholds.derive import flag_records_for_work_from_home_rules
-from cishouseholds.derive import flag_records_for_work_location_null
-from cishouseholds.derive import flag_records_for_work_location_student
-from cishouseholds.derive import regex_match_result
 from cishouseholds.edit import clean_job_description_string
 from cishouseholds.edit import update_column_values_from_map
+from cishouseholds.expressions import any_column_not_null
 from cishouseholds.impute import fill_backwards_work_status_v2
 from cishouseholds.impute import fill_forward_from_last_change_marked_subset
 from cishouseholds.impute import fill_forward_only_to_nulls
+from cishouseholds.merge import left_join_keep_right
 from cishouseholds.pyspark_utils import get_or_create_spark_session
-from cishouseholds.regex.regex_patterns import at_school_pattern
-from cishouseholds.regex.regex_patterns import at_university_pattern
-from cishouseholds.regex.regex_patterns import childcare_pattern
-from cishouseholds.regex.regex_patterns import furloughed_pattern
-from cishouseholds.regex.regex_patterns import in_college_or_further_education_pattern
-from cishouseholds.regex.regex_patterns import not_working_pattern
-from cishouseholds.regex.regex_patterns import retired_regex_pattern
-from cishouseholds.regex.regex_patterns import self_employed_regex
-from cishouseholds.regex.regex_patterns import work_from_home_pattern
+
+from typing import Optional
+from cishouseholds.pipeline.lookup_and_regex_transformations import process_healthcare_regex, reclassify_work_variables
+
 
 # this wall feed in data from the joined healthcare regex
 
 
-def job_transformations(df: DataFrame):
+def job_transformations(df: DataFrame, soc_lookup_df:DataFrame, job_lookup_df: Optional[DataFrame]=None):
     """apply all transformations in order related to a persons vocation."""
-    df = fill_forwards(df).custom_checkpoint()
-    df = data_dependent_transformations(df).custom_checkpoint()
-    return df
+    df = preprocessing(df)
+    job_lookup_df = create_job_lookup(df,soc_lookup_df=soc_lookup_df,lookup_df=job_lookup_df).custom_checkpoint()
+    df = left_join_keep_right(left_df=df,right_df=job_lookup_df,join_on_columns=["work_main_job_title", "work_main_job_role"])
+    df = fill_forwards_and_backwards(df).custom_checkpoint()
+    df = data_dependent_derivations(df).custom_checkpoint()
+    return df, job_lookup_df
 
 
 def preprocessing(df: DataFrame):
@@ -73,7 +43,57 @@ def preprocessing(df: DataFrame):
     return df
 
 
-def fill_forwards(df: DataFrame):
+def get_unprocessed_rows(df:DataFrame, lookup_df:Optional[DataFrame] = None):
+    """"""
+    join_on_columns = ["work_main_job_title", "work_main_job_role"]
+    df = df.filter(any_column_not_null(join_on_columns))
+    if lookup_df is None:
+        df_to_process = df.dropDuplicates(join_on_columns)
+    else:
+        non_derived_rows = df.join(lookup_df, on=join_on_columns, how="leftanti")
+        df_to_process = non_derived_rows.dropDuplicates(join_on_columns)
+    partitions = int(get_or_create_spark_session().sparkContext.getConf().get("spark.sql.shuffle.partitions"))
+    partitions = int(partitions / 2)
+    df_to_process = df_to_process.repartition(partitions)
+    print(
+        f"     - creating regex lookup table from {df_to_process.count()} rows. This may take some time ... "
+    )  # functional
+    return df_to_process
+
+
+def create_job_lookup(df:DataFrame, soc_lookup_df:DataFrame, lookup_df:Optional[DataFrame]=None):
+    df_to_process = get_unprocessed_rows(df,lookup_df)
+    df_to_process = process_healthcare_regex(df_to_process)
+    df_to_process = reclassify_work_variables(df_to_process)
+    df_to_process = df_to_process.join(soc_lookup_df,on=["work_main_job_title", "work_main_job_role"],how="left")
+
+    processed_df = df_to_process.select(
+        "work_main_job_title",
+        "work_main_job_role",
+
+        #from healthcare
+        "work_direct_contact_patients_or_clients",
+        "work_social_care_area",
+        "work_health_care_area",
+        "work_health_care_patient_facing",
+        "work_patient_facing_clean",
+        "work_social_care",
+        "works_health_care",
+        "work_nursing_or_residential_care_home",
+        "regex_derived_job_sector",
+
+        #from reclassify
+        "work_status_v0",
+        "work_status_v1",
+        "work_status_v2",
+        "work_location"
+    )
+    if lookup_df is not None:
+        return lookup_df.unionByName(processed_df)
+    return processed_df
+
+
+def fill_forwards_and_backwards(df: DataFrame):
     df = fill_forward_from_last_change_marked_subset(
         df=df,
         fill_forward_columns=[
@@ -94,12 +114,6 @@ def fill_forwards(df: DataFrame):
         dateset_version_column="survey_response_dataset_major_version",
         minimum_dateset_version=2,
     )
-    return df
-
-
-def data_dependent_transformations(df: DataFrame) -> DataFrame:
-    """Apply transformations that require all data to have been previously filled over rows."""
-    df = df.withColumn("work_main_job_title_and_role", F.concat_ws(" ", "work_main_job_title", "work_main_job_role"))
     df = fill_backwards_work_status_v2(
         df=df,
         date="visit_datetime",
@@ -114,6 +128,25 @@ def data_dependent_transformations(df: DataFrame) -> DataFrame:
             "Attending university (including if temporarily absent)",
         ],
     )
+    df = fill_forward_only_to_nulls(
+        df,
+        id="participant_id",
+        date="visit_datetime",
+        list_fill_forward=[
+            "work_status_v0",
+            "work_status_v1",
+            "work_status_v2",
+            "work_location",
+            "work_not_from_home_days_per_week",
+        ],
+    )
+    return df
+
+
+def data_dependent_derivations(df: DataFrame) -> DataFrame:
+    """Apply transformations that require all data to have been previously filled over rows."""
+    df = df.withColumn("work_main_job_title_and_role", F.concat_ws(" ", "work_main_job_title", "work_main_job_role"))
+
     df = assign_work_status_group(df, "work_status_group", "work_status_v0")
 
     window = Window.partitionBy("participant_id")
@@ -124,362 +157,6 @@ def data_dependent_transformations(df: DataFrame) -> DataFrame:
     df = df.withColumn(
         "patient_facing_over_20_percent", F.when(patient_facing_percentage >= 0.2, "Yes").otherwise("No")
     )
-    return df
-
-
-def reclassify_work_variables(df: DataFrame, drop_original_variables: bool = True) -> DataFrame:
-    """
-    Reclassify work-related variables based on rules & regex patterns
-
-    Parameters
-    ----------
-    df
-        The dataframe containing the work-status related variables we want to edit
-    drop_original_variables
-        Set this to False if you want to retain the original variables so you can compare
-        before & after edits.
-    """
-    spark_session = get_or_create_spark_session()
-    # Work from Home
-    working_from_home_regex_hit = regex_match_result(
-        columns_to_check_in=["work_main_job_title", "work_main_job_role"],
-        positive_regex_pattern=work_from_home_pattern.positive_regex_pattern,
-        negative_regex_pattern=work_from_home_pattern.negative_regex_pattern,
-    )
-    # Rule_id: 1000
-    update_work_location = flag_records_for_work_from_home_rules() & working_from_home_regex_hit
-
-    # Furlough
-    furlough_regex_hit = regex_match_result(
-        columns_to_check_in=["work_main_job_title", "work_main_job_role"],
-        positive_regex_pattern=furloughed_pattern.positive_regex_pattern,
-        negative_regex_pattern=furloughed_pattern.negative_regex_pattern,
-    )
-
-    # Rule_id: 2000
-    update_work_status_furlough_v0 = furlough_regex_hit & flag_records_for_furlough_rules_v0()
-    # Rule_id: 2001
-    update_work_status_furlough_v1_a = furlough_regex_hit & flag_records_for_furlough_rules_v1_a()
-    # Rule_id: 2002
-    update_work_status_furlough_v1_b = furlough_regex_hit & flag_records_for_furlough_rules_v1_b()
-    # Rule_id: 2003
-    update_work_status_furlough_v2_a = furlough_regex_hit & flag_records_for_furlough_rules_v2_a()
-    # Rule_id: 2004
-    update_work_status_furlough_v2_b = furlough_regex_hit & flag_records_for_furlough_rules_v2_b()
-
-    # Self-Employed
-    self_employed_regex_hit = regex_match_result(
-        columns_to_check_in=["work_main_job_title", "work_main_job_role"],
-        positive_regex_pattern=self_employed_regex.positive_regex_pattern,
-        negative_regex_pattern=self_employed_regex.negative_regex_pattern,
-    )
-
-    # Rule_id: 3000
-    update_work_status_self_employed_v0 = self_employed_regex_hit & flag_records_for_self_employed_rules_v0()
-    # Rule_id: 3001
-    update_work_status_self_employed_v1_a = self_employed_regex_hit & flag_records_for_self_employed_rules_v1_a()
-    # Rule_id: 3002
-    update_work_status_self_employed_v1_b = self_employed_regex_hit & flag_records_for_self_employed_rules_v1_b()
-    # Rule_id: 3003
-    update_work_status_self_employed_v2_a = self_employed_regex_hit & flag_records_for_self_employed_rules_v2_a()
-    # Rule_id: 3004
-    update_work_status_self_employed_v2_b = self_employed_regex_hit & flag_records_for_self_employed_rules_v2_b()
-
-    # Retired
-    retired_regex_hit = regex_match_result(
-        columns_to_check_in=["work_main_job_title", "work_main_job_role"],
-        positive_regex_pattern=retired_regex_pattern.positive_regex_pattern,
-        negative_regex_pattern=retired_regex_pattern.negative_regex_pattern,
-    )
-
-    # Rule_id: 4000, 4001, 4002
-    update_work_status_retired = retired_regex_hit | flag_records_for_retired_rules()
-
-    # Not-working
-    not_working_regex_hit = (
-        regex_match_result(
-            columns_to_check_in=["work_main_job_title", "work_main_job_role"],
-            positive_regex_pattern=not_working_pattern.positive_regex_pattern,
-            negative_regex_pattern=not_working_pattern.negative_regex_pattern,
-        )
-        & ~working_from_home_regex_hit  # type: ignore
-    )
-
-    # Rule_id: 5000
-    update_work_status_not_working_v0 = not_working_regex_hit & flag_records_for_not_working_rules_v0()
-    # Rule_id: 5001
-    update_work_status_not_working_v1_a = not_working_regex_hit & flag_records_for_not_working_rules_v1_a()
-    # Rule_id: 5002
-    update_work_status_not_working_v1_b = not_working_regex_hit & flag_records_for_not_working_rules_v1_b()
-    # Rule_id: 5003
-    update_work_status_not_working_v2_a = not_working_regex_hit & flag_records_for_not_working_rules_v2_a()
-    # Rule_id: 5004
-    update_work_status_not_working_v2_b = not_working_regex_hit & flag_records_for_not_working_rules_v2_b()
-
-    # School/Student
-    school_regex_hit = regex_match_result(
-        columns_to_check_in=["work_main_job_title", "work_main_job_role"],
-        positive_regex_pattern=at_school_pattern.positive_regex_pattern,
-        negative_regex_pattern=at_school_pattern.negative_regex_pattern,
-    )
-
-    college_regex_hit = regex_match_result(
-        columns_to_check_in=["work_main_job_title", "work_main_job_role"],
-        positive_regex_pattern=in_college_or_further_education_pattern.positive_regex_pattern,
-        negative_regex_pattern=in_college_or_further_education_pattern.negative_regex_pattern,
-    )
-
-    university_regex_hit = regex_match_result(
-        columns_to_check_in=["work_main_job_title", "work_main_job_role"],
-        positive_regex_pattern=at_university_pattern.positive_regex_pattern,
-        negative_regex_pattern=at_university_pattern.negative_regex_pattern,
-    )
-
-    # Childcare
-    childcare_regex_hit = regex_match_result(
-        columns_to_check_in=["work_main_job_title", "work_main_job_role"],
-        positive_regex_pattern=childcare_pattern.positive_regex_pattern,
-        negative_regex_pattern=childcare_pattern.negative_regex_pattern,
-    )
-
-    age_under_16 = F.col("age_at_visit") < F.lit(16)
-    age_over_four = F.col("age_at_visit") > F.lit(4)
-
-    # Rule_id: 6000
-    update_work_status_student_v0 = (
-        (school_regex_hit & flag_records_for_school_v2_rules())
-        | (university_regex_hit & flag_records_for_uni_v0_rules())
-        | (college_regex_hit & flag_records_for_college_v0_rules())
-        | (age_over_four & age_under_16)
-    )
-
-    # Rule_id: 6001
-    update_work_status_student_v0_a = (childcare_regex_hit & flag_records_for_childcare_v0_rules()) | (
-        school_regex_hit & flag_records_for_childcare_v0_rules()
-    )
-
-    # Rule_id: 6002
-    update_work_status_student_v1_a = (
-        (school_regex_hit & flag_records_for_school_v2_rules())
-        | (university_regex_hit & flag_records_for_uni_v1_rules())
-        | (college_regex_hit & flag_records_for_college_v1_rules())
-        | (age_over_four & age_under_16)
-    )
-
-    # Rule_id: 6003
-    update_work_status_student_v1_c = (childcare_regex_hit & flag_records_for_childcare_v1_rules()) | (
-        school_regex_hit & flag_records_for_childcare_v1_rules()
-    )
-
-    # Rule_id: 6004
-    update_work_status_student_v2_e = (childcare_regex_hit & flag_records_for_childcare_v2_b_rules()) | (
-        school_regex_hit & flag_records_for_childcare_v2_b_rules()
-    )
-
-    # Rule_id: 6005
-    update_work_status_student_v2_a = (school_regex_hit & flag_records_for_school_v2_rules()) | (
-        age_over_four & age_under_16
-    )
-
-    # Rule_id: 6006
-    update_work_status_student_v2_b = college_regex_hit & flag_records_for_college_v2_rules()
-
-    # Rule_id: 6007
-    update_work_status_student_v2_c = university_regex_hit & flag_records_for_uni_v2_rules()
-
-    # Rule_id: 6008
-    update_work_location_general = flag_records_for_work_location_null() | flag_records_for_work_location_student()
-
-    # Please note the order of *_edited columns, these must come before the in-place updates
-
-    # first start by taking a copy of the original work variables
-    _df = (
-        df.withColumn("work_location_original", F.col("work_location"))
-        .withColumn("work_status_v0_original", F.col("work_status_v0"))
-        .withColumn("work_status_v1_original", F.col("work_status_v1"))
-        .withColumn("work_status_v2_original", F.col("work_status_v2"))
-        .withColumn(
-            "work_location",
-            F.when(update_work_location, F.lit("Working from home")).otherwise(F.col("work_location")),
-        )
-        .withColumn(
-            "work_status_v0",
-            F.when(update_work_status_self_employed_v0, F.lit("Self-employed")).otherwise(F.col("work_status_v0")),
-        )
-        .withColumn(
-            "work_status_v1",
-            F.when(update_work_status_self_employed_v1_a, F.lit("Self-employed and currently working")).otherwise(
-                F.col("work_status_v1")
-            ),
-        )
-        .withColumn(
-            "work_status_v1",
-            F.when(update_work_status_self_employed_v1_b, F.lit("Self-employed and currently not working")).otherwise(
-                F.col("work_status_v1")
-            ),
-        )
-        .withColumn(
-            "work_status_v2",
-            F.when(
-                update_work_status_self_employed_v2_a,
-                F.lit("Self-employed and currently working"),
-            ).otherwise(F.col("work_status_v2")),
-        )
-        .withColumn(
-            "work_status_v2",
-            F.when(update_work_status_self_employed_v2_b, F.lit("Self-employed and currently not working")).otherwise(
-                F.col("work_status_v2")
-            ),
-        )
-    )
-
-    _df2 = spark_session.createDataFrame(_df.rdd, schema=_df.schema)  # breaks lineage to avoid Java OOM Error
-
-    _df3 = (
-        _df2.withColumn(
-            "work_status_v0",
-            F.when(update_work_status_student_v0 | update_work_status_student_v0_a, F.lit("Student")).otherwise(
-                F.col("work_status_v0")
-            ),
-        )
-        .withColumn(
-            "work_status_v1",
-            F.when(update_work_status_student_v1_a, F.lit("5y and older in full-time education")).otherwise(
-                F.col("work_status_v1")
-            ),
-        )
-        .withColumn(
-            "work_status_v1",
-            F.when(update_work_status_student_v1_c, F.lit("Child under 5y attending child care")).otherwise(
-                F.col("work_status_v1")
-            ),
-        )
-        .withColumn(
-            "work_status_v2",
-            F.when(update_work_status_student_v2_a, F.lit("4-5y and older at school/home-school")).otherwise(
-                F.col("work_status_v2")
-            ),
-        )
-        .withColumn(
-            "work_status_v2",
-            F.when(
-                update_work_status_student_v2_b, F.lit("Attending college or FE (including if temporarily absent)")
-            ).otherwise(F.col("work_status_v2")),
-        )
-        .withColumn(
-            "work_status_v2",
-            F.when(
-                update_work_status_student_v2_c, F.lit("Attending university (including if temporarily absent)")
-            ).otherwise(F.col("work_status_v2")),
-        )
-        .withColumn(
-            "work_status_v2",
-            F.when(update_work_status_student_v2_e, F.lit("Child under 4-5y attending child care")).otherwise(
-                F.col("work_status_v2")
-            ),
-        )
-        .withColumn(
-            "work_status_v0",
-            F.when(
-                update_work_status_retired, F.lit("Not working (unemployed, retired, long-term sick etc.)")
-            ).otherwise(F.col("work_status_v0")),
-        )
-        .withColumn(
-            "work_status_v1",
-            F.when(update_work_status_retired, F.lit("Retired")).otherwise(F.col("work_status_v1")),
-        )
-        .withColumn(
-            "work_status_v2",
-            F.when(update_work_status_retired, F.lit("Retired")).otherwise(F.col("work_status_v2")),
-        )
-    )
-
-    _df4 = spark_session.createDataFrame(_df3.rdd, schema=_df3.schema)  # breaks lineage to avoid Java OOM Error
-
-    _df5 = (
-        _df4.withColumn(
-            "work_status_v0",
-            F.when(
-                update_work_status_not_working_v0, F.lit("Not working (unemployed, retired, long-term sick etc.)")
-            ).otherwise(F.col("work_status_v0")),
-        )
-        .withColumn(
-            "work_status_v1",
-            F.when(update_work_status_not_working_v1_a, F.lit("Employed and currently not working")).otherwise(
-                F.col("work_status_v1")
-            ),
-        )
-        .withColumn(
-            "work_status_v1",
-            F.when(update_work_status_not_working_v1_b, F.lit("Self-employed and currently not working")).otherwise(
-                F.col("work_status_v1")
-            ),
-        )
-        .withColumn(
-            "work_status_v2",
-            F.when(update_work_status_not_working_v2_a, F.lit("Employed and currently not working")).otherwise(
-                F.col("work_status_v2")
-            ),
-        )
-        .withColumn(
-            "work_status_v2",
-            F.when(update_work_status_not_working_v2_b, F.lit("Self-employed and currently not working")).otherwise(
-                F.col("work_status_v2")
-            ),
-        )
-        .withColumn(
-            "work_status_v0",
-            F.when(update_work_status_furlough_v0, F.lit("Furloughed (temporarily not working)")).otherwise(
-                F.col("work_status_v0")
-            ),
-        )
-        .withColumn(
-            "work_status_v1",
-            F.when(update_work_status_furlough_v1_a, F.lit("Employed and currently not working")).otherwise(
-                F.col("work_status_v1")
-            ),
-        )
-        .withColumn(
-            "work_status_v1",
-            F.when(update_work_status_furlough_v1_b, F.lit("Self-employed and currently not working")).otherwise(
-                F.col("work_status_v1")
-            ),
-        )
-        .withColumn(
-            "work_status_v2",
-            F.when(update_work_status_furlough_v2_a, F.lit("Employed and currently not working")).otherwise(
-                F.col("work_status_v2")
-            ),
-        )
-        .withColumn(
-            "work_status_v2",
-            F.when(update_work_status_furlough_v2_b, F.lit("Self-employed and currently not working")).otherwise(
-                F.col("work_status_v2")
-            ),
-        )
-        .withColumn(
-            "work_location",
-            F.when(
-                update_work_location_general,
-                F.lit("Not applicable, not currently working"),
-            ).otherwise(F.col("work_location")),
-        )
-    )
-
-    if drop_original_variables:
-        # replace original versions with their cleaned versions
-        _df5 = _df5.drop(
-            "work_location_original",
-            "work_status_v0_original",
-            "work_status_v1_original",
-            "work_status_v2_original",
-        )
-
-    return _df5
-
-
-def assign_work_classifications(df: DataFrame):
-    """"""
     df = assign_work_patient_facing_now(
         df,
         column_name_to_assign="work_patient_facing_now",
@@ -499,16 +176,5 @@ def assign_work_classifications(df: DataFrame):
     #     "work_status_v0",
     #     ["Furloughed (temporarily not working)", "Not working (unemployed, retired, long-term sick etc.)", "Student"],
     # )
-    df = reclassify_work_variables(df, drop_original_variables=False)
-    df = fill_forward_only_to_nulls(
-        df,
-        id="participant_id",
-        date="visit_datetime",
-        list_fill_forward=[
-            "work_status_v0",
-            "work_status_v1",
-            "work_status_v2",
-            "work_location",
-            "work_not_from_home_days_per_week",
-        ],
-    )
+    return df
+
